+++ lib/src/signals/port.dart
@@ -30,4 +30,6 @@ class Port extends Logic {
       throw InvalidPortNameException(name);
     }
   }
+
+  //TODO: should we deprecate Port for Logic.port, then typedef it?
 }